+++ Tests/MMExtensionTests.m
@@ -201,6 +201,15 @@
     );
 }
 
+- (void)testAsterisksInsideWord
+{
+    MMAssertExtendedMarkdownEqualsHTML(
+        MMMarkdownExtensionsUnderscoresInWords,
+        @"t*es*t",
+        @"<p>t<em>es</em>t</p>"
+    );
+}
+
 
 #pragma mark - MMMarkdownExtensionsAutolinkedURLs
 
@@ -471,6 +480,36 @@
         @"<pre><code class=\"objc\">hello\nworld\n"
         "</code></pre>\n"
     );
+}
+
+- (void)testFencedCodeBlockWithSpaceBeforeLanguage
+{
+    MMAssertExtendedMarkdownEqualsHTML(
+        MMMarkdownExtensionsFencedCodeBlocks,
+        @"``` objc\nhello\nworld\n```",
+        @"<pre><code class=\"objc\">hello\nworld\n"
+        "</code></pre>\n"
+    );
+}
+
+- (void)testFencedCodeBlockWithSpaceAfterLanguage
+{
+    MMAssertExtendedMarkdownEqualsHTML(
+        MMMarkdownExtensionsFencedCodeBlocks,
+        @"```objc \nhello\nworld\n```",
+        @"<pre><code class=\"objc\">hello\nworld\n"
+        "</code></pre>\n"
+    );
+}
+
+- (void)testFencedCodeBlockWithSpaceInLanguageName
+{
+    MMAssertExtendedMarkdownEqualsHTML(
+        MMMarkdownExtensionsFencedCodeBlocks,
+        @"```a b\nhello\nworld\n```",
+        @"<p>```a b\nhello\nworld</p>\n<pre><code></code></pre>\n"
+    );
+    
 }
 
 - (void)testFencedCodeBlockInsideBlockquote